+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Improved `table` view footer rendering to handle and display errors that occur during cell rendering, instead of silently failing with no output
 - Added `footer` option for `cols` in text-based `table` view
 - Added `headerWhen` and `footerWhen` options for text-based `table` view
+- Improved performace for long list rendering with `ViewModel#view.renderList()` method
 - Updated the `ViewModel#view.renderList()` method to accept the `limit` as a separate argument outside of the `options` object:  
   `renderList(container, itemConfig, data, context, offset, limit, options)`.